+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - SphereInteractor component
 - CollisionableUnityEvent component
 
+### Changed
+- Rename UnityEventInteractable into InteractableUnityEvent
+
 ## [1.0.0] - 2025-10-30
 ### Added
 - UnityEventInteractable component