+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Rename UnityEventInteractable into InteractableUnityEvent
 - Refact IInteractable to implement ICollisionable
 
+### Removed
+- Interactor component. Use any AbstractInteractor implementation (BoxInteractor or SphereInteractor)
+
 ## [1.0.0] - 2025-10-30
 ### Added
 - UnityEventInteractable component