+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.0] - 2026-03-10
 ### Added
 - ICollisionable interface
 - BoxInteractor component
@@ -31,5 +32,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/0.1.0...main
-[0.1.0]: https://github.com/1mbitshorde/InteractableSystem/tree/0.1.0/
+[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/2.0.0...main
+[2.0.0]: https://github.com/1mbitshorde/InteractableSystem/tree/2.0.0/
+[1.0.0]: https://github.com/1mbitshorde/InteractableSystem/tree/1.0.0/